1) update RPi2 to send messages to server 2) remove RPiConsole
