+++ RPi2/RPi Server API/DeviceData.cs
@@ -7,7 +7,6 @@ using System.Text;
 using System.Threading;
 using System.Diagnostics;
 using Windows.UI.Core;
-//using Windows.System.
 
 namespace RPi2.RPi_Server_API
 {
@@ -17,26 +16,14 @@ namespace RPi2.RPi_Server_API
     public sealed class CDeviceData
     {
         #region Fields
-
+        
         private static volatile CDeviceData m_instance;
         private static object syncRoot = new object();
-        private static DeviceClient m_client = null;
+        private MessageConverter messageConvert;
+        private CDeviceMessagesSendReceive deviceMessagesSendReceive;
 
-        private static readonly string iotHubUri = "smartchair-iothub.azure-devices.net";
         private static readonly string deviceId = "SmartChair01";
         private static readonly string deviceKey = "Sgerz/a7KV2M8/kJ+As5XH5u/o9fJtIIuDsQZYpLsGU=";
-
-        public static DeviceClient Client
-        {
-            get
-            {
-                if (m_client == null)
-                {
-                    m_client = DeviceClient.Create(iotHubUri, new DeviceAuthenticationWithRegistrySymmetricKey(deviceId, deviceKey));
-                }
-                return m_client;
-            }
-        }
 
         internal Windows.UI.Xaml.Controls.TextBlock guiDebugging = null;
 
@@ -46,7 +33,8 @@ namespace RPi2.RPi_Server_API
 
         private CDeviceData()
         {
-            
+            messageConvert = MessageConverter.Instance;
+            deviceMessagesSendReceive = new CDeviceMessagesSendReceive(deviceId, deviceKey);
         }
 
         #endregion
@@ -96,23 +84,38 @@ namespace RPi2.RPi_Server_API
         /// RPi sends data to Azure server, including device id, a list of normalized measurements and an end of sample timestamp
         /// keeps sending until returned success
         /// </summary>
-        public async void RPiServer_newDataSample(System.DateTime timestamp)
+        public void RPiServer_newDataSample(System.DateTime timestamp)
         {
-            List<int> pressureList = new List<int>();
-            foreach (KeyValuePair<EChairPart, Dictionary<EChairPartArea, int>> part in Data)
+            string messageString = createMessageStringFromData(timestamp);
+            deviceMessagesSendReceive.sendMessageToServerAsync(messageString);
+            reportMessageSent(messageString);
+        }
+
+        private string createMessageStringFromData(DateTime timestamp)
+        {
+            int[] pressureArr = aggregateDataToArray();
+            Datapoint dataPoint = new Datapoint(deviceId, timestamp, pressureArr);
+            string messageString = messageConvert.encode(EMessageId.RpiServer_Datapoint, dataPoint);
+
+            return messageString;
+        }
+
+        private int[] aggregateDataToArray()
+        {
+            List<int> sensorDataList = new List<int>();
+            foreach (var chairPart in Data)
             {
-                foreach (KeyValuePair<EChairPartArea, int> partArea in part.Value)
+                foreach (var partArea in chairPart.Value)
                 {
-                    pressureList.Add(partArea.Value);
+                    sensorDataList.Add(partArea.Value);
                 }
             }
 
-            Datapoint dataPoint = new Datapoint(deviceId, timestamp, pressureList.ToArray());
-            Message<Datapoint> messageStruct = new Message<Datapoint>(EMessageId.RpiServer_Datapoint, dataPoint);
-            string messageString = JsonConvert.SerializeObject(messageStruct);
-            Message message = new Message(Encoding.ASCII.GetBytes(messageString));
-            await Client.SendEventAsync(message);
+            return sensorDataList.ToArray();
+        }
 
+        private async void reportMessageSent(string messageString)
+        {
             if (guiDebugging != null)
             {
                 // call dispacher to update gui from another thread:
