+++ smartchair.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+VisualStudioVersion = 14.0.25123.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Server", "Server\Server.csproj", "{634D6169-5762-4EA8-BFC2-282A125E64F2}"
 EndProject
@@ -12,8 +12,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RPiSimulator", "RPiSimulator\RPiSimulator.csproj", "{C14AF19B-EECF-414B-B74A-712186EE4D6B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClientSimulator", "ClientSimulator\ClientSimulator.csproj", "{E9859EB1-8286-4957-A902-A9425DC97E18}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RPiConsole", "RPiConsole\RPiConsole.csproj", "{B859B6D9-1156-4F6A-B311-A29D39981331}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RPi2", "RPi2\RPi2.csproj", "{2DC5A524-C6DD-45BE-A4E1-2BEAC937BDCF}"
 EndProject
@@ -112,21 +110,6 @@ Global
 		{E9859EB1-8286-4957-A902-A9425DC97E18}.Release|x64.Build.0 = Release|Any CPU
 		{E9859EB1-8286-4957-A902-A9425DC97E18}.Release|x86.ActiveCfg = Release|Any CPU
 		{E9859EB1-8286-4957-A902-A9425DC97E18}.Release|x86.Build.0 = Release|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Debug|ARM.Build.0 = Debug|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Debug|x64.Build.0 = Debug|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Debug|x86.Build.0 = Debug|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Release|ARM.ActiveCfg = Release|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Release|ARM.Build.0 = Release|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Release|x64.ActiveCfg = Release|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Release|x64.Build.0 = Release|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Release|x86.ActiveCfg = Release|Any CPU
-		{B859B6D9-1156-4F6A-B311-A29D39981331}.Release|x86.Build.0 = Release|Any CPU
 		{2DC5A524-C6DD-45BE-A4E1-2BEAC937BDCF}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{2DC5A524-C6DD-45BE-A4E1-2BEAC937BDCF}.Debug|ARM.ActiveCfg = Debug|ARM
 		{2DC5A524-C6DD-45BE-A4E1-2BEAC937BDCF}.Debug|ARM.Build.0 = Debug|ARM
